Rename chrome-remote-debug-protocol
