+++ packages/chrome-remote-debugging-protocol/bin/build-bootstrap.js
