+++ packages/chrome-remote-debugging-protocol/bootstrap.js
