+++ packages/devtools-local-toolbox/package.json
@@ -25,7 +25,7 @@
     "amd-loader": "0.0.5",
     "body-parser": "^1.15.0",
     "check-node-version": "^1.1.2",
-    "chrome-remote-debug-protocol": "^0.0.3",
+    "chrome-remote-debugging-protocol": "^0.0.3",
     "co": "=4.6.0",
     "css-loader": "^0.25.0",
     "devtools-config": "^0.0.3",
